Migrate the NPBD model to the source code
