+++ src/ClimaOceanBiogeochemistry.jl
@@ -1,5 +1,5 @@
 module ClimaOceanBiogeochemistry
 
-greet() = print("Hello World!")
+include("nutrients_plankton_bacteria_detritus.jl")
 
 end # module ClimaOceanBiogeochemistry
